Move build_key_collection onto KeyCollection as a named constructor
build_key_collection took a KeyCollection's own inputs (a constraint
system and an inout split) and returned a KeyCollection, so it was a
named constructor in disguise, living as a free function in a sibling
module instead of on the type it builds. This moves it to
KeyCollection::build, matching the ShiftProver/KeySegment/build_g
methods already established in this module.

Pure rename: no behavior change. Updates every call site across
binius-prover (including its benches and tests), binius-m4-prover, and
binius-examples, and moves the function's test module from builder.rs
to collection.rs alongside its new home.
