+++ crates/examples/src/cli.rs
@@ -829,7 +829,7 @@ where
 		// Save KeyCollection if requested
 		if let Some(path) = key_collection_path.as_deref() {
 			let key_collection_scope = tracing::info_span!("Building key collection").entered();
-			let key_collection = binius_prover::protocols::shift::build_key_collection(
+			let key_collection = binius_prover::protocols::shift::KeyCollection::build(
 				cs,
 				binius_core::constraint_system::InoutSegment::Public,
 			);
